+++ benchmark/Chart.hs
@@ -1,18 +1,13 @@
-{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Main where
 
-import Control.Exception (handle, catch, SomeException, ErrorCall(..))
+import Control.Exception (catch, ErrorCall(..))
 import Control.Monad.Trans.State
 import Control.Monad.Trans.Maybe
 import Data.Char (toLower)
-import Data.Function (on, (&))
 import Data.List
-import Data.List.Split
-import Data.Maybe (mapMaybe)
-import Data.Ord (comparing)
 import System.Environment (getArgs)
 import Control.Monad.IO.Class (liftIO)
 import Control.Monad (mzero)
@@ -24,32 +19,27 @@ import BenchShow
 ------------------------------------------------------------------------------
 
 data BenchType
-    = Linear
-    | LinearAsync
-    | LinearRate
-    | NestedConcurrent
-    | FileIO
-    | Concurrent
-    | Parallel
-    | Adaptive
-    | Compare String
+    = Compare String
     | Standard String
     deriving Show
 
 data Options = Options
     { genGraphs :: Bool
-    , benchType :: BenchType
+    , benchType :: Maybe BenchType
     } deriving Show
 
-defaultOptions = Options False Linear
+defaultOptions :: Options
+defaultOptions = Options False Nothing
 
+setGenGraphs :: Monad m => Bool -> StateT (a, Options) m ()
 setGenGraphs val = do
     (args, opts) <- get
     put (args, opts { genGraphs = val })
 
+setBenchType :: Monad m => BenchType -> StateT (a, Options) m ()
 setBenchType val = do
     (args, opts) <- get
-    put (args, opts { benchType = val })
+    put (args, opts { benchType = Just val })
 
 -- Like the shell "shift" to shift the command line arguments
 shift :: StateT ([String], Options) (MaybeT IO) (Maybe String)
@@ -63,14 +53,6 @@ parseBench :: StateT ([String], Options) (MaybeT IO) ()
 parseBench = do
     x <- shift
     case x of
-        Just "linear" -> setBenchType Linear
-        Just "linear-async" -> setBenchType LinearAsync
-        Just "linear-rate" -> setBenchType LinearRate
-        Just "nested-concurrent" -> setBenchType NestedConcurrent
-        Just "fileio" -> setBenchType FileIO
-        Just "concurrent" -> setBenchType Concurrent
-        Just "parallel" -> setBenchType Parallel
-        Just "adaptive" -> setBenchType Adaptive
         Just str | "_cmp" `isSuffixOf` str -> setBenchType (Compare str)
         Just str -> setBenchType (Standard str)
         Nothing -> do
@@ -101,101 +83,9 @@ parseOptions = do
             Just opt -> parseOpt opt >> parseLoop
             Nothing -> return ()
 
+ignoringErr :: IO () -> IO ()
 ignoringErr a = catch a (\(ErrorCall err :: ErrorCall) ->
     putStrLn $ "Failed with error:\n" <> err <> "\nSkipping.")
-
-------------------------------------------------------------------------------
--- Linear composition charts
-------------------------------------------------------------------------------
-
-makeLinearGraphs :: Config -> String -> IO ()
-makeLinearGraphs cfg@Config{..} inputFile = do
-    ignoringErr $ graph inputFile "generation" $ cfg
-        { title = (++) <$> title <*> Just " generation"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/generation/"
-        }
-
-    ignoringErr $ graph inputFile "elimination" $ cfg
-        { title = (++) <$> title <*> Just " Elimination"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/elimination/"
-        }
-
-    ignoringErr $ graph inputFile "transformation-zip" $ cfg
-        { title = (++) <$> title <*> Just " Transformation & Zip"
-        , classifyBenchmark = \b ->
-                if    "serially/transformation/" `isPrefixOf` b
-                   || "serially/zipping" `isPrefixOf` b
-                then Just ("Streamly", last $ splitOn "/" b)
-                else Nothing
-        }
-
-    ignoringErr $ graph inputFile "filtering" $ cfg
-        { title = (++) <$> title <*> Just " Filtering"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/filtering/"
-        }
-
-    ignoringErr $ graph inputFile "transformationX4" $ cfg
-        { title = (++) <$> title <*> Just " Transformation x 4"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/transformationX4/"
-        }
-
-    ignoringErr $ graph inputFile "filteringX4"
-        $ cfg
-        { title = (++) <$> title <*> Just " Filtering x 4"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/filteringX4/"
-        }
-
-    ignoringErr $ graph inputFile "mixedX4"
-        $ cfg
-        { title = (++) <$> title <*> Just " Mixed x 4"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/mixedX4/"
-        }
-
-    ignoringErr $ graph inputFile "iterated"
-        $ cfg
-        { title = Just "iterate 10,000 times over 10 elems"
-        , classifyBenchmark =
-            fmap ("Streamly",) . stripPrefix "serially/iterated/"
-        }
-
-------------------------------------------------------------------------------
--- Stream type based comparison charts
-------------------------------------------------------------------------------
-
-makeStreamComparisonGraphs :: String -> [String] -> Config -> String -> IO ()
-makeStreamComparisonGraphs outputPrefix benchPrefixes cfg inputFile =
-    ignoringErr $ graph inputFile outputPrefix $ cfg
-        { presentation = Groups Absolute
-        , classifyBenchmark = classifyNested
-        , selectGroups = \gs ->
-            groupBy ((==) `on` snd) gs
-            & fmap (\xs -> mapMaybe (\x -> (x,) <$> lookup x xs) benchPrefixes)
-            & concat
-        }
-
-    where
-
-    classifyNested b
-        | "serially/" `isPrefixOf` b =
-            ("serially",) <$> stripPrefix "serially/" b
-        | "asyncly/" `isPrefixOf` b =
-            ("asyncly",) <$> stripPrefix "asyncly/" b
-        | "wAsyncly/" `isPrefixOf` b =
-            ("wAsyncly",) <$> stripPrefix "wAsyncly/" b
-        | "aheadly/" `isPrefixOf` b =
-            ("aheadly",) <$> stripPrefix "aheadly/" b
-        | "parallely/" `isPrefixOf` b =
-            ("parallely",) <$> stripPrefix "parallely/" b
-        | otherwise = Nothing
-
-linearAsyncPrefixes = ["asyncly", "wAsyncly", "aheadly", "parallely"]
-nestedBenchPrefixes = ["serially"] ++ linearAsyncPrefixes
 
 ------------------------------------------------------------------------------
 -- Generic
@@ -209,6 +99,7 @@ makeGraphs name cfg@Config{..} inputFile =
 -- Arrays
 ------------------------------------------------------------------------------
 
+showComparisons :: Options -> Config -> FilePath -> FilePath -> IO ()
 showComparisons Options{..} cfg inp out =
     let cfg1 = cfg { classifyBenchmark = classifyComparison }
      in if genGraphs
@@ -222,7 +113,7 @@ showComparisons Options{..} cfg inp out =
 
     dropComponent = dropWhile (== '/') . dropWhile (/= '/')
 
-    classifyComparison b = Just $
+    classifyComparison b = Just
         ( takeWhile (/= '/') b
         , dropComponent b
         )
@@ -242,6 +133,13 @@ selectBench f =
       (sortOn snd)
       $ f (ColumnIndex 1) (Just PercentDiff)
 
+benchShow ::
+       Options
+    -> Config
+    -> (Config -> String -> IO ())
+    -> String
+    -> FilePath
+    -> IO ()
 benchShow Options{..} cfg func inp out =
     if genGraphs
     then func cfg {outputDir = Just out} inp
@@ -256,7 +154,7 @@ main = do
                 ( flip elem ["time" , "mean"
                             , "maxrss", "cputime"
                             ]
-                . map toLower
+                . fmap toLower
                 )
             }
     res <- parseOptions
@@ -267,52 +165,16 @@ main = do
             return ()
         Just opts@Options{..} ->
             case benchType of
-                Linear -> benchShow opts cfg
-                            { title = Just "Linear" }
-                            makeLinearGraphs
-                            "charts/linear/results.csv"
-                            "charts/linear"
-                LinearRate -> benchShow opts cfg
-                            { title = Just "Linear Rate" }
-                            (makeGraphs "linear-rate")
-                            "charts/linear-rate/results.csv"
-                            "charts/linear-rate"
-                LinearAsync -> benchShow opts cfg
-                            { title = Just "Linear Async" }
-                            (makeStreamComparisonGraphs "linear-async" linearAsyncPrefixes)
-                            "charts/linear-async/results.csv"
-                            "charts/linear-async"
-                NestedConcurrent -> benchShow opts cfg
-                            { title = Just "Nested concurrent loops" }
-                            (makeStreamComparisonGraphs "nested-concurrent" nestedBenchPrefixes)
-                            "charts/nested-concurrent/results.csv"
-                            "charts/nested-concurrent"
-                FileIO -> benchShow opts cfg
-                            { title = Just "File IO" }
-                            (makeGraphs "fileIO")
-                            "charts/fileio/results.csv"
-                            "charts/fileio"
-                Concurrent -> benchShow opts cfg
-                            { title = Just "Concurrent Ops" }
-                            (makeGraphs "Concurrent")
-                            "charts/concurrent/results.csv"
-                            "charts/concurrent"
-                Parallel -> benchShow opts cfg
-                            { title = Just "Parallel" }
-                            (makeGraphs "parallel")
-                            "charts/parallel/results.csv"
-                            "charts/parallel"
-                Adaptive -> benchShow opts cfg
-                            { title = Just "Adaptive" }
-                            (makeGraphs "adaptive")
-                            "charts/adaptive/results.csv"
-                            "charts/adaptive"
-                Compare str -> showComparisons opts cfg
-                            { title = Just $ str }
-                            ("charts/" ++ str ++ "/results.csv")
-                            ("charts/" ++ str)
-                Standard str -> benchShow opts cfg
-                            { title = Just str }
-                            (makeGraphs str)
-                            ("charts/" ++ str ++ "/results.csv")
-                            ("charts/" ++ str)
+                Just (Compare str) ->
+                    showComparisons opts cfg
+                        { title = Just str }
+                        ("charts/" ++ str ++ "/results.csv")
+                        ("charts/" ++ str)
+                Just (Standard str) ->
+                    benchShow opts cfg
+                        { title = Just str }
+                        (makeGraphs str)
+                        ("charts/" ++ str ++ "/results.csv")
+                        ("charts/" ++ str)
+                Nothing ->
+                    error "Please specify a benchmark using --benchmark."
