--- benchmark/Streamly/Benchmark/Prelude/LinearRate.hs
@@ -1,24 +1,0 @@
--- |
--- Module      : Main
--- Copyright   : (c) 2018 Harendra Kumar
---
--- License     : BSD3
--- Maintainer  : streamly@composewell.com
-
-import Streamly.Benchmark.Common
-import Streamly.Benchmark.Prelude
-
-import Gauge
-
-main :: IO ()
-main = do
-    (value, cfg, benches) <- parseCLIOpts defaultStreamSize
-    value `seq` runMode (mode cfg) cfg benches (allBenchmarks value)
-
-    where
-
-    allBenchmarks value =
-        concat
-            [ o_1_space_async_avgRate value
-            , o_1_space_ahead_avgRate value
-            ]
